update main program
